+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/ai_guard/models/CompetitorsAction.java
@@ -1,0 +1,35 @@
+package cloud.pangeacyber.pangea.ai_guard.models;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonValue;
+
+public enum CompetitorsAction {
+	REPORT("report"),
+	BLOCK("block");
+
+	private final String value;
+
+	CompetitorsAction(final String value) {
+		this.value = value;
+	}
+
+	@JsonValue
+	public String getValue() {
+		return value;
+	}
+
+	@Override
+	public String toString() {
+		return String.valueOf(value);
+	}
+
+	@JsonCreator
+	public static CompetitorsAction fromValue(final String value) {
+		for (var x : CompetitorsAction.values()) {
+			if (x.value.equals(value)) {
+				return x;
+			}
+		}
+		throw new IllegalArgumentException("Unexpected value '" + value + "'");
+	}
+}