+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/ai_guard/models/TextGuardDetector.java
@@ -3,13 +3,14 @@ package cloud.pangeacyber.pangea.ai_guard.models;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.Data;
+import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
+import lombok.Value;
 
-@Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE, force = true)
+@Value
 public final class TextGuardDetector<T> {
 
 	boolean detected;