+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/ai_guard/requests/TextGuardRequest.java
@@ -1,6 +1,8 @@
 package cloud.pangeacyber.pangea.ai_guard.requests;
 
 import cloud.pangeacyber.pangea.BaseRequest;
+import cloud.pangeacyber.pangea.ai_guard.models.LogFields;
+import cloud.pangeacyber.pangea.ai_guard.models.Overrides;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
@@ -33,13 +35,6 @@ public final class TextGuardRequest<TMessages> extends BaseRequest {
 	TMessages messages;
 
 	/**
-	 * Structured full llm payload data to be scanned by AI Guard for PII,
-	 * sensitive data, malicious content, and other data types defined by the
-	 * configuration. Supports processing up to 10KB of JSON text.
-	 */
-	TMessages llmInput;
-
-	/**
 	 * Recipe key of a configuration of data types and settings defined in the
 	 * Pangea User Console. It specifies the rules that are to be applied to the
 	 * text, such as defang malicious URLs.
@@ -51,4 +46,9 @@ public final class TextGuardRequest<TMessages> extends BaseRequest {
 	 * data
 	 */
 	boolean debug;
+
+	Overrides overrides;
+
+	/** Additional fields to include in activity log. */
+	LogFields logFields;
 }